+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "ionic-docs" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.0.0]
+
+- Removed webview - It now opens as a native vscode markdown preview 
+- Added window progress
+- Changed icons
+- Removed open to website on right click.
+
 ## [0.0.4]
 
 - Initial release